refactor: move bitcoin config the bitcoin repo
